+++ lib/src/storages/test_result_storage.dart
@@ -132,7 +132,7 @@ Future<void> updateTestResultAttachmentsAsync(
     });
 
 @internal
-Future<void> updateTestResultLinksAsync(final List<Link> links) async =>
+Future<void> updateTestResultLinksAsync(final Iterable<Link> links) async =>
     await _lock.synchronized(() => _testResults.update(_getTestId(), (value) {
           value.links.addAll(links);
 
@@ -196,7 +196,7 @@ String _getTestId() {
   return testId;
 }
 
-void _updateCurrentStepAttachments(final AttachmentPutModel attachment) async {
+void _updateCurrentStepAttachments(final AttachmentPutModel attachment) {
   final currentStep = _getCurrentStep();
   currentStep?.attachments.add(attachment);
 }