+++ test/step_manager_test.dart
@@ -4,9 +4,17 @@ import 'package:adapters_flutter/adapters_flutter.dart';
 
 void main() {
   group('step manager', () {
-    test('run step', () {
+    test('step', () {
       expect(() async {
         await step('', () {});
+      }, returnsNormally);
+    });
+
+    test('step with value', () {
+      expect(() async {
+        await step('', () {
+          return 0;
+        });
       }, returnsNormally);
     });
   });